requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
django-allauth
dj-rest-auth
google-api-python-client